Ran rake db:migrate

+++ db/migrate/20161109192417_create_reimbursements.rb
@@ -1,7 +1,7 @@
 class CreateReimbursements < ActiveRecord::Migration
   def change
     create_table :reimbursements do |t|
-      t.double :total
+      t.decimal :total
       t.string :description
       t.date :request_date
       t.string :status
